seprated event and venue in different apps
