+++ src/templates/base.html
@@ -35,13 +35,13 @@
                         </a>
                     </li>
                     <li class="" >
-                        <a href="{% url 'workshop:event_index' %}">
+                        <a href="{% url 'workshop:index' %}">
                             <span class="glyphicon glyphicon-time" aria-hidden="true"></span>&nbsp;
                             Events
                         </a>
                     </li>
                     <li class="" >
-                        <a href="{% url 'workshop:venue_index' %}">
+                        <a href="{% url 'venue:index' %}">
                             <span class="glyphicon glyphicon-calendar" aria-hidden="true"></span>&nbsp;
                             Venue
                         </a>
