+++ src/templates/workshop/event_list.html
@@ -3,19 +3,22 @@
 {% block title %} Events List | {{ block.super }} {% endblock %}
 
 {% block content %}
+	{% if user.is_authenticated %}
 	<div class="row">
 		<div class="col-md-offset-8 col-md-4">
-			<a class="btn btn-primary btn-block" href="{% url 'workshop:event_create'%}">Add New Event</a>
+			<a class="btn btn-primary btn-block" href="{% url 'workshop:create'%}">Add New Event</a>
 		</div>
 	</div>
 	<br>
+	{% endif %}
+	
 	<div class="row">
 	{% if events %}
 		{% for event in events %}
 			<div class="col-md-4">
 				<div class="panel panel-default">
 					<div class="panel-heading">
-						<a href="{% url 'workshop:event_detail' event.id %}">
+						<a href="{% url 'workshop:detail' event.id %}">
 						{{ event.activityName }}
 						</a>
 					</div>
