+++ src/venueapp/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import Event
+from .models import Venue
 
-admin.site.register(Event)
-
+admin.site.register(Venue)
 # Register your models here.
