+++ src/venueapp/forms.py
@@ -1,0 +1,7 @@
+from django import forms
+from .models import Venue
+
+class VenueForm(forms.ModelForm):
+	class Meta:
+		model = Venue
+		fields = '__all__'
